+++ lib/supabase/.gitignore
