+++ lib/supabase/auth.dart
