+++ lib/supabase/scouting/questions.dart
@@ -1,9 +1,43 @@
 import 'dart:convert';
 
+import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
-import '../../database.dart';
-import '../../model/scouting.dart';
+import '../database.dart';
+import 'scouting.dart';
+
+part 'questions.freezed.dart';
+part 'questions.g.dart';
+
+@immutable
+@Freezed(fromJson: true, toJson: true)
+sealed class QuestionNode with _$QuestionNode {
+  const factory QuestionNode.group({
+    required String id,
+    required int season,
+    required ScoutingCategory category,
+    String? prompt,
+    String? parentId,
+    int? index,
+  }) = QuestionGroup;
+
+  const factory QuestionNode.question({
+    required String id,
+    required int season,
+    required ScoutingCategory category,
+    required String prompt,
+    required String parentId,
+    required int index,
+    required DataType dataType,
+    required Map<String, dynamic> config,
+  }) = Question;
+
+  factory QuestionNode.fromJson(Map<String, dynamic> json) {
+    return json['data_type'] == null
+        ? QuestionGroup.fromJson(json)
+        : Question.fromJson(json);
+  }
+}
 
 class QuestionsRepository {
   final Cache<(int, ScoutingCategory), List<QuestionNode>> _questionsCache;
